requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
slowapi==0.1.9